services/job-matcher/requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai==1.59.9
structlog==24.4.0
requests==2.32.3
pytz==2024.2
werkzeug==3.1.3
jinja2==3.1.4
python-dotenv==1.0.1
gunicorn==23.0.0
supabase==2.9.1
httpx==0.27.2